Added movement controls and shooting mechanics
The ship moves with the WASD keys and can shoot at a forward direction on the screen by clicking or holding down the mouse button. Changing the mouse position while firing also changes the trajectory of the projectiles. Done that by mistake but it looks kinda cool...

I sort of was able to rotate the space ship to aim in the direction of the mouse but it would fall out of alignment after a bit of movement. Currently doesn't work but the roll is partially implemented.