Perfected movement and added First Person View
Movement is now glides and the ship has velocity for a better feel. The ship can now overstep the boundaries so it doesnt just feel like it hits an invisible wall.

First person view implemented. Made a variable that tracks the ship's up vector which is applied to the camera when the 'v' key is pressed. The camera is offset above and behind the ship. It updates dynamically as the up vector is changed. Movement also takes the vector into account. so turning left and right while upside down isn't inverted